Data migration pull request
This is a small program use that output the data of the first ten users who submitted the most and submitted the most in a day into an Excel table. Make sure import panda openpyxl in python and already have a app.db(When running this website, the program will automatically generate this database).